face-recognition-service/requirements.txt
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
flask-cors==4.0.0
Pillow==12.1.0
numpy==1.26.4
firebase-admin==6.5.0
requests==2.32.3
